# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GensanWorks API
  version: 1.0.0
  description: Core public and authenticated API surface for jobs, applications, messaging, and analytics.
servers:
  - url: http://localhost:3000
paths:
...
  /api/jobs/{id}/apply:
    post:
      summary: Apply to a job (jobseeker auth required)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coverLetter:
                  type: string
      responses:
        "200":
          description: Application accepted
        "401":
          description: Unauthorized
        "409":
          description: Duplicate application
...
  /api/messages:
    get:
      summary: List conversations/messages
      responses:
        "200":
          description: Messages payload
        "401":
          description: Unauthorized
    post:
      summary: Send message
      responses:
        "200":
          description: Message sent
        "401":
          description: Unauthorized
...